+++ components/auth/index.ts
@@ -1,5 +1,13 @@
 import GoogleLoginButton from "./GoogleLoginButton";
 import KakaoLoginButton from "./KakaoLoginButton";
+import OauthLoginGroup from "./OauthLoginGroup";
+import TermsAndConditionsModal from "./TermsAndConditionsModal";
 import LoginForm from "./LoginForm";
 
-export { GoogleLoginButton, KakaoLoginButton, LoginForm };
+export {
+  GoogleLoginButton,
+  KakaoLoginButton,
+  LoginForm,
+  TermsAndConditionsModal,
+  OauthLoginGroup,
+};
